validation and all methods complete

+++ HealthAPI/HealthAPI.DTOs/Patients/PatientDTO.cs
@@ -15,11 +15,11 @@
 
         public string Email { get; set; }
 
-        public int? Age { get; set; }
+        public double? Age { get; set; }
 
-        public int? Height { get; set; }
+        public double? Height { get; set; }
 
-        public int? Weight { get; set; }
+        public double? Weight { get; set; }
 
         public string Gender { get; set; }
