+++ HealthAPI/HealthAPI.Data/Model/Patient.cs
@@ -13,11 +13,11 @@
 
         public string Email { get; set; }
 
-        public int? Age { get; set; }
+        public double? Age { get; set; }
 
-        public int? Height { get; set; }
+        public double? Height { get; set; }
 
-        public int? Weight { get; set; }
+        public double? Weight { get; set; }
 
         public string Insurance { get; set; }
